[jupyter] Add ipynb support
It is a bit awkward at the moment. There are two major parts to remember
for ipynb posts that is different than other posts:

1. The name of the file is used as the name in the post listing. Not
precisely sure why the title is not picked up here.
2. The first cell in the ipynb has a few restrictions: the first cell
must be raw, must contain yaml that maps to Map String String, and must
include a tag called "meta" to be recognized.

For now I will just live with these tradeoffs since there are not that
many ipynb posts (for now...).